Refactor ServiceFlowPage and new-services.store
